feat: move header navigation to hook
